# internal/infra/http/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Gym API
  version: 1.0.0
...
paths:
...
  /groups:
    post:
      tags: [Groups]
      summary: Create a group
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '201':
          description: Created group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    get:
      tags: [Groups]
      summary: List the current user's groups
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupListItem'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /groups/join:
    post:
      tags: [Groups]
      summary: Join a group by invite code
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinGroupRequest'
      responses:
        '200':
          description: Joined group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Already a member
  /groups/{id}:
    get:
      tags: [Groups]
      summary: Group detail (header with weekly scores)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Group detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /groups/{id}/ranking:
    get:
      tags: [Groups]
      summary: Weekly ranking of group members
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Ranking entries ordered by points desc
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RankingEntry'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /groups/{id}/feed:
    get:
      tags: [Groups]
      summary: Group feed (cursor-paginated)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Feed page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /groups/{id}/sessions/{sessionId}:
    get:
      tags: [Groups]
      summary: Session detail as a social post (group-scoped)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Session post header with reactions and comment count
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /groups/{id}/sessions/{sessionId}/reaction:
    put:
      tags: [Groups]
      summary: Set or toggle the viewer's emoji reaction on a session
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetReactionRequest'
      responses:
        '200':
          description: Updated session detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Groups]
      summary: Remove the viewer's reaction on a session
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Updated session detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /groups/{id}/sessions/{sessionId}/comments:
    get:
      tags: [Groups]
      summary: List a session's comments (cursor-paginated)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Comment page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Groups]
      summary: Comment on a session
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCommentRequest'
      responses:
        '201':
          description: Created comment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /groups/{id}/sessions/{sessionId}/comments/{commentId}:
    delete:
      tags: [Groups]
      summary: Delete your own comment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: commentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /groups/{id}/cover:
    patch:
      tags: [Groups]
      summary: Set the group cover image (owner only, multipart)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: Group with cover_url set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /groups/{id}/leave:
    delete:
      tags: [Groups]
      summary: Leave a group
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Left the group
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string

    # ── Pagination ────────────────────────────────────────────────────────────
    CursorMeta:
      type: object
      properties:
        next_cursor:
          type: string
          nullable: true
          description: Opaque token to pass as `?cursor=` to retrieve the next page. Null when there are no more pages.
        has_more:
          type: boolean
...
    CreateGroupRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
    JoinGroupRequest:
      type: object
      required: [invite_code]
      properties:
        invite_code:
          type: string
    Group:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        cover_url:
          type: string
          nullable: true
        invite_code:
          type: string
        owner_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
    GroupListItem:
      allOf:
        - $ref: '#/components/schemas/Group'
        - type: object
          properties:
            member_count:
              type: integer
            member_avatars:
              type: array
              items:
                type: string
              description: Up to 3 member avatar URLs for the list avatar stack.
    GroupDetail:
      allOf:
        - $ref: '#/components/schemas/Group'
        - type: object
          properties:
            top_score:
              type: integer
            top_name:
              type: string
            top_avatar_url:
              type: string
              nullable: true
            my_score:
              type: integer
            member_count:
              type: integer
    RankingEntry:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        name:
          type: string
        avatar_url:
          type: string
          nullable: true
        points:
          type: integer
    FeedItem:
      type: object
      properties:
        session_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
        workout_name:
          type: string
        photo_url:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        reaction_count:
          type: integer
        comment_count:
          type: integer
        my_reaction:
          type: string
          nullable: true
          description: The viewer's emoji on this session in this group, or null.
        top_emoji:
          type: string
          nullable: true
          description: Most-used emoji on this session (tie broken by earliest), or null.
    FeedPage:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/FeedItem'
        cursor:
          $ref: '#/components/schemas/CursorMeta'
    ReactionCount:
      type: object
      properties:
        emoji:
          type: string
        count:
          type: integer
    SessionDetail:
      type: object
      properties:
        session_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
        avatar_url:
          type: string
          nullable: true
        workout_name:
          type: string
        photo_url:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        reactions:
          type: array
          items:
            $ref: '#/components/schemas/ReactionCount'
        my_reaction:
          type: string
          nullable: true
        comment_count:
          type: integer
    CommentItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
        avatar_url:
          type: string
          nullable: true
        body:
          type: string
        created_at:
          type: string
          format: date-time
        is_mine:
          type: boolean
    CommentPage:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/CommentItem'
        cursor:
          $ref: '#/components/schemas/CursorMeta'
    SetReactionRequest:
      type: object
      required: [emoji]
      properties:
        emoji:
          type: string
    AddCommentRequest:
      type: object
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 500
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...